address arch findings: move cache/health logic to utils, fix auth 5xx handling and stale fetch race
- Extract Redis probe and response assembly from cache_status, cache_info, and
  health_check route handlers into get_cache_status(), get_cache_info(), and
  get_health_status() in utils.py; consolidate duplicate _redact_redis_url and
  _fmt_ttl helpers into utils and remove them from routes
- Replace raw fetch() fallback in DevoFeedback with apiGet() for consistent auth headers
- validateAuth now throws for non-401/403 errors (e.g. 502) so useRequireAuth
  treats transient outages as session-preserving rather than forcing logout
- Add AbortController to DevoFeedbackEdit.loadFeedback to cancel stale responses
  when dateStr or section changes mid-flight
- Update tests to reflect new patch targets and 5xx throw semantics
